requirements.txt
#python==3.10.11
google-generativeai==0.8.3
Pillow
python-dotenv
streamlit
